Use distinct queues for user events (panning, zooming) and programmatic events (e.g. changed bounding box after GPS-Update) to never miss a user event.
There was a problem with the single queue solution: if too many GPS updates occured (each second or more often) the event queue was blocked from start event until the end of deceleration (fastest is 1 sec), so the queue was actually blocked all the time, events triggered by the user were ignored most of the time. Now there are two queues, one for programmtic events and one for user events. User events take precedence. If a user event is in progress, no programmatic event is queued. But if a programmatic event is running, the user event is queued nonetheless. But you can ask the RMMapView’s property programmaticEventInProgress, if a programmatic event is happening right now, so it’s up to your event handling to decide if and how the event should be handled.